refactor(gym):파트너 헬스장 분리 /GYMMATE-309
