+++ src/main/java/org/helloworld/gymmate/domain/gym/gymInfo/dto/request/RegisterGymRequest.java
@@ -7,7 +7,6 @@ import org.helloworld.gymmate.domain.gym.gymProduct.dto.GymProductRequest;
 public record RegisterGymRequest(
 	Long gymId,
 	GymInfoRequest gymInfoRequest,
-	List<GymProductRequest> gymProductRequest,
-	List<String> imageUrl
+	List<GymProductRequest> gymProductRequest
 ) {
 }
