+++ src/main/java/org/helloworld/gymmate/domain/gym/gymProduct/mapper/GymProductMapper.java
@@ -1,8 +1,8 @@
 package org.helloworld.gymmate.domain.gym.gymProduct.mapper;
 
-import org.helloworld.gymmate.domain.gym.gymInfo.entity.PartnerGym;
 import org.helloworld.gymmate.domain.gym.gymProduct.dto.GymProductRequest;
 import org.helloworld.gymmate.domain.gym.gymProduct.entity.GymProduct;
+import org.helloworld.gymmate.domain.gym.partnerGym.entity.PartnerGym;
 
 import jakarta.validation.Valid;
