+++ src/main/java/org/helloworld/gymmate/domain/gym/gymProduct/repository/GymProductRepository.java
@@ -2,8 +2,8 @@ package org.helloworld.gymmate.domain.gym.gymProduct.repository;
 
 import java.util.List;
 
-import org.helloworld.gymmate.domain.gym.gymInfo.entity.PartnerGym;
 import org.helloworld.gymmate.domain.gym.gymProduct.entity.GymProduct;
+import org.helloworld.gymmate.domain.gym.partnerGym.entity.PartnerGym;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
