+++ src/main/java/org/helloworld/gymmate/domain/gym/partnerGym/mapper/PartnerGymMapper.java
@@ -1,4 +1,4 @@
-package org.helloworld.gymmate.domain.gym.gymInfo.mapper;
+package org.helloworld.gymmate.domain.gym.partnerGym.mapper;
 
 import java.util.List;
 
@@ -9,8 +9,8 @@ import org.helloworld.gymmate.domain.gym.gymInfo.dto.response.GymProductResponse
 import org.helloworld.gymmate.domain.gym.gymInfo.dto.response.PartnerGymDetailResponse;
 import org.helloworld.gymmate.domain.gym.gymInfo.entity.Gym;
 import org.helloworld.gymmate.domain.gym.gymInfo.entity.GymImage;
-import org.helloworld.gymmate.domain.gym.gymInfo.entity.PartnerGym;
 import org.helloworld.gymmate.domain.gym.gymProduct.entity.GymProduct;
+import org.helloworld.gymmate.domain.gym.partnerGym.entity.PartnerGym;
 
 public class PartnerGymMapper {
